[fixCI]: use workaround to make the CI passed
